design: 동료 이벤트 컴포넌트 마크업 및 스타일 수정 / 파일명 오타 수정 (#82)
